requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.9
django-multiselectfield==0.1.8
django-toolbelt==0.0.1
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.5
static3==0.7.0
